Return all the response keys to the model
Pydantic will ignore unknown field names when creating the model. This
lets us return all the keys from the response that we get, so that we
can add new fields to the model in the future without modifying the
request code. Unless, of course, the response from the external API
needs some more munging. But nothing to be done about that.