Add authorization requirement to submit papers
Users have to log in with a GitHub account to be able to submit a paper.
This will serve to hopefully reduce the amount of spam we get. To
prevent CORS errors, the login cannot be triggered by JavaScript, that
is the login cannot be triggered by clicking the submit button. Intead a
GitHub login link is added to the submit page.

This also refactors the approval page to use a similar mechanism, where
you have to click a link to log in to view that page.