Refactor cookie handling and auth
The cookie is now read in a dependency function and a new Pydantic model
is returned if the cookie is available. Each function can then decide
what it wants to do with that model, rather than having a centralized
decision function.

Also adds a logout link that deletes the cookie and requires clicking
the login button again.