Add theme to the site
Copies theme details from the API docs and the main website. There
aren't any links to the application in theme yet, deferring that until
deploy time.

Removes the JSON encoding extension that's not needed anymore.

The 302 status code seems to be needed to have the page refresh when
login details are available.